requirements.txt
pytest-playwright==0.4.4
playwright==1.55.0
